CRUD tickets

+++ src/main/java/com/mfrakhman/etick/ticket/entity/Ticket.java
@@ -6,6 +6,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -26,7 +27,7 @@ public class Ticket {
     private Integer totalTicket;
 
     @Column(name = "price", nullable = false)
-    private Integer price;
+    private BigDecimal price;
 
     // Many tickets belong to one event
     @ManyToOne
